+++ sidebar/src/components/UI/molecules/search_guide/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Card from '@material-ui/core/Card';
 import useStyle from './style';
-import useMousePosition from '../../../../lib/useMousePosition';
+import useMousePosition from '../../../../hooks/useMousePosition';
 
 const SearchGuide = ({ isHover }) => {
   const { posX, posY } = useMousePosition();
